Extract local logistics to their own file.
