+++ .gitignore
@@ -10,7 +10,7 @@ Source/KolonyTools/bin/Debug/Mono.Cecil.dll
 Source/KolonyTools/bin/Debug/UnityEngine.dll
 Source/KolonyTools/bin/Debug/KolonyTools.pdb
 Source/KolonyTools/bin/Debug/TDx.TDxInput.dll
-Source/KolonyTools/obj/Debug/DesignTimeResolveAssemblyReferencesInput.cache
+Source/KolonyTools/obj/
 Source/KolonyTools/obj/Debug/KolonyTools.pdb
 Source/KolonyTools/obj/Debug/KolonyTools.csproj.FileListAbsolute.txt
 Source/KolonyTools/obj/Debug/DesignTimeResolveAssemblyReferences.cache
