+++ Source/KolonyTools/KolonyTools/KolonyTools.csproj
@@ -84,6 +84,7 @@
     <Compile Include="Kolonization\KolonizationMonitor.cs" />
     <Compile Include="Kolonization\KolonizationPersistance.cs" />
     <Compile Include="Kolonization\KolonizationScenario.cs" />
+    <Compile Include="Kolonization/ManualLocalLogistics.cs" />
     <Compile Include="MKSModule.cs" />
     <Compile Include="OrbitalLogistics\MKSMainGui.cs" />
     <Compile Include="OrbitalLogistics\MKSTransferCreateView.cs" />
@@ -106,7 +107,7 @@
   <ItemGroup />
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <PropertyGroup>
-    <PostBuildEvent>"D:\Games\pdb2mdb\pdb2mdb.exe" "$(TargetFileName)"</PostBuildEvent>
+    <PostBuildEvent Condition="'$(OS)' != 'Unix'">"D:\Games\pdb2mdb\pdb2mdb.exe" "$(TargetFileName)"</PostBuildEvent>
   </PropertyGroup>
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it.
 Other similar extension points exist, see Microsoft.Common.targets.
